requirements.txt
streamlit
Pillow
fpdf
PyPDF2
